+++ languagetool-core/src/main/java/org/languagetool/tools/RuleMatchesAsJsonSerializer.java
@@ -20,7 +20,6 @@ package org.languagetool.tools;
 
 import com.fasterxml.jackson.core.JsonFactory;
 import com.fasterxml.jackson.core.JsonGenerator;
-import org.languagetool.Experimental;
 import org.languagetool.JLanguageTool;
 import org.languagetool.Language;
 import org.languagetool.rules.Category;
@@ -46,8 +45,8 @@ public class RuleMatchesAsJsonSerializer {
 
   private final JsonFactory factory = new JsonFactory();
   
-  public String ruleMatchesToJson(List<RuleMatch> matches, String text, int contextSize, Language lang) {
-    return ruleMatchesToJson(matches, new ArrayList<>(), text, contextSize, lang, null);
+  public String ruleMatchesToJson(List<RuleMatch> matches, String text, int contextSize, Language lang, Language detectedLang) {
+    return ruleMatchesToJson(matches, new ArrayList<>(), text, contextSize, lang, detectedLang, null);
   }
 
   /**
@@ -56,7 +55,7 @@ public class RuleMatchesAsJsonSerializer {
    * @since 3.7
    */
   public String ruleMatchesToJson(List<RuleMatch> matches, List<RuleMatch> hiddenMatches, String text, int contextSize,
-                                  Language lang, String incompleteResultsReason) {
+                                  Language lang, Language detectedLang, String incompleteResultsReason) {
     ContextTools contextTools = new ContextTools();
     contextTools.setEscapeHtml(false);
     contextTools.setContextSize(contextSize);
@@ -68,7 +67,7 @@ public class RuleMatchesAsJsonSerializer {
         g.writeStartObject();
         writeSoftwareSection(g);
         writeWarningsSection(g, incompleteResultsReason);
-        writeLanguageSection(g, lang);
+        writeLanguageSection(g, lang, detectedLang);
         writeMatchesSection("matches", g, matches, text, contextTools);
         if (hiddenMatches != null && hiddenMatches.size() > 0) {
           writeMatchesSection("hiddenMatches", g, hiddenMatches, text, contextTools);
@@ -106,10 +105,16 @@ public class RuleMatchesAsJsonSerializer {
     g.writeEndObject();
   }
 
-  private void writeLanguageSection(JsonGenerator g, Language lang) throws IOException {
+  private void writeLanguageSection(JsonGenerator g, Language lang, Language detectedLang) throws IOException {
     g.writeObjectFieldStart("language");
     g.writeStringField("name", lang.getName());
     g.writeStringField("code", lang.getShortCodeWithCountryAndVariant());
+    if (detectedLang != null) {
+      g.writeObjectFieldStart("detectedLanguage");
+      g.writeStringField("name", detectedLang.getName());
+      g.writeStringField("code", detectedLang.getShortCodeWithCountryAndVariant());
+      g.writeEndObject();
+    }
     g.writeEndObject();
   }