+++ languagetool-core/src/test/java/org/languagetool/tools/RuleMatchesAsJsonSerializerTest.java
@@ -47,7 +47,7 @@ public class RuleMatchesAsJsonSerializerTest {
 
   @Test
   public void testJson() {
-    String json = serializer.ruleMatchesToJson(matches, "This is an text.", 5, Languages.getLanguageForShortCode("xx-XX"));
+    String json = serializer.ruleMatchesToJson(matches, "This is an text.", 5, Languages.getLanguageForShortCode("xx-XX"), null);
     // Software:
     assertContains("\"LanguageTool\"", json);
     assertContains(JLanguageTool.VERSION, json);
@@ -71,13 +71,13 @@ public class RuleMatchesAsJsonSerializerTest {
 
   @Test
   public void testJsonWithUnixLinebreak() {
-    String json = serializer.ruleMatchesToJson(matches, "This\nis an text.", 5, Languages.getLanguageForShortCode("xx-XX"));
+    String json = serializer.ruleMatchesToJson(matches, "This\nis an text.", 5, Languages.getLanguageForShortCode("xx-XX"), null);
     assertTrue(json.contains("This is ..."));  // got filtered out by ContextTools
   }
   
   @Test
   public void testJsonWithWindowsLinebreak() {
-    String json = serializer.ruleMatchesToJson(matches, "This\ris an text.", 5, Languages.getLanguageForShortCode("xx-XX"));
+    String json = serializer.ruleMatchesToJson(matches, "This\ris an text.", 5, Languages.getLanguageForShortCode("xx-XX"), null);
     assertTrue(json.contains("This\\ris ..."));
   }